backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.0
scikit-learn==1.5.2
shap==0.46.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
google-generativeai==0.8.0
python-multipart==0.0.12
